updated PartialMetricInstrument to implement MetricInstrument and updated boolean method name to isEnabledByDefault
